server/requirements.txt
# FastAPI and ASGI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Supabase
supabase>=2.27.0

# OpenAI
openai>=1.3.0

# LangChain and AI
langchain>=0.1.0
langchain-openai>=0.1.0
llama-index>=0.14.0

# PDF Processing
pypdf>=6.0.0
pdfplumber>=0.11.0
unstructured>=0.18.0

# Data Processing
pandas>=2.2.0
openpyxl>=3.1.0

# Environment
python-dotenv>=1.0.0
